Move profiling.h one level up
