+++ include/concore2full/detail/core_types.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "concore2full/detail/profiling.h"
+#include "concore2full/profiling.h"
 
 #include <context_core_api.h>
