+++ include/concore2full/thread_pool.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "concore2full/detail/profiling.h"
+#include "concore2full/profiling.h"
 
 #include <thread>
 #include <vector>
